# docs/swagger.yaml
definitions:
...
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
  request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.MenuRequest:
    properties:
      created_by:
        type: string
      icon:
        type: string
      name:
        type: string
      ord:
        type: integer
      parent_id:
        type: string
      updated_by:
        type: string
      url:
        type: string
    required:
    - name
    - ord
    - url
    type: object
...
  request.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Api Service boilerplate dari golang gin dan gorm
  termsOfService: http://swagger.io/terms/
  title: My Admin Service Api
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Handle login user and return the received data.
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Login User
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Handle register user and return the received data.
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Register User
      tags:
      - Auth
  /menus:
    get:
      consumes:
      - application/json
      description: REST API Menu
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get list menu (requires authentication)
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: REST API menu
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.MenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Insert menu
      tags:
      - Menus
  /menus/{id}:
    delete:
      consumes:
      - application/json
      description: REST API menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Delete menu
      tags:
      - Menus
    get:
      consumes:
      - application/json
      description: REST API menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get one menu
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: REST API menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        type: string
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.MenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Update menu
      tags:
      - Menus
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"